requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
fastapi-clerk-auth
pydantic
